requirements.txt
# ============================================
# Multi-Modal RAG QA System - Requirements
# ============================================

# Core Dependencies
python-dotenv>=1.0.0

# ============================================
# PDF & Document Processing
# ============================================
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
pdf2image>=1.16.3
pypdf>=3.17.0

# ============================================
# Image Processing & OCR
# ============================================
Pillow>=10.1.0
opencv-python>=4.8.1
pytesseract>=0.3.10
easyocr>=1.7.0

# ============================================
# Video Processing
# ============================================
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# ============================================
# Audio Processing
# ============================================
openai-whisper>=20231117
pydub>=0.25.1
librosa>=0.10.1

# ============================================
# Vector Database
# ============================================
chromadb>=0.4.18
qdrant-client>=1.7.0
faiss-cpu>=1.7.4

# ============================================
# Embeddings & Models
# ============================================
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.1.0
torchvision>=0.16.0
huggingface-hub>=0.19.0

# ============================================
# Multi-Modal Embeddings (CLIP)
# ============================================
open-clip-torch>=2.23.0
ftfy>=6.1.1
regex>=2023.10.3

# ============================================
# Retrieval & Ranking
# ============================================
rank-bm25>=0.2.2
scikit-learn>=1.3.2

# ============================================
# LLM Integration
# ============================================
openai>=1.3.0
langchain>=0.0.340
langchain-community>=0.0.1
langchain-huggingface>=0.0.1
langchain-core>=0.1.0
groq>=0.4.1

# ============================================
# Evaluation Frameworks
# ============================================
ragas>=0.1.0
nltk>=3.8.1
rouge-score>=0.1.2
bert-score>=0.3.13

# ============================================
# Data Processing
# ============================================
pandas>=2.1.3
numpy>=1.26.2
datasets>=2.15.0

# ============================================
# API & Web Framework
# ============================================
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
httpx>=0.25.1
streamlit>=1.28.0

# ============================================
# Configuration Management
# ============================================
pyyaml>=6.0.1
omegaconf>=2.3.0
hydra-core>=1.3.2

# ============================================
# Utilities
# ============================================
tqdm>=4.66.1
loguru>=0.7.2
rich>=13.7.0
tenacity>=8.2.3

# ============================================
# Testing
# ============================================
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0

# ============================================
# Jupyter & Notebooks
# ============================================
jupyter>=1.0.0
ipywidgets>=8.1.1
notebook>=7.0.6
jupyterlab>=4.0.9

# ============================================
# Visualization
# ============================================
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0

# ============================================
# Documentation
# ============================================
mkdocs>=1.5.3
mkdocs-material>=9.4.14